refactor: update imports to use types from 'types/settings' instead of 'types/index.d'
