+++ src/__test__/mocks/network/msw/handle-get-signatures.ts
@@ -7,7 +7,7 @@ import { JSNS } from '@zextras/carbonio-shell-ui';
 
 import { createSoapAPIInterceptor } from '@test-utils/network/msw/create-api-interceptor';
 import { GetSignaturesRequest, GetSignaturesResponse } from 'api/get-signatures-soap-api';
-import { SignItemType } from 'types';
+import { SignItemType } from 'types/settings';
 
 export const handleGetSignaturesRequest = (signatures: Array<SignItemType>): void => {
 	createSoapAPIInterceptor<GetSignaturesRequest, GetSignaturesResponse>('GetSignatures', {
