+++ src/views/settings/signature-settings.tsx
@@ -24,7 +24,7 @@ import { reject, concat, map } from 'lodash';
 import { v4 as uuidv4 } from 'uuid';
 
 import { NO_SIGNATURE_ID, NO_SIGNATURE_LABEL } from 'helpers/signatures';
-import type { SignatureSettingsPropsType, SignItemType } from 'types/index.d';
+import { SignatureSettingsPropsType, SignItemType } from 'types/settings';
 import SelectIdentitySignature from 'views/settings/components/select-identity-signature';
 import { getFonts, getFontSizesOptions } from 'views/settings/components/utils';
 import { ListOld } from 'views/settings/list-old';
